+++ packages/quarks/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [1.12.0] - 2026-09-18
+
+### Changed
+
+- Removed automatic generic `gluon` and `quark` classes from Quark output;
+  component-specific classes remain intact.
+
 ## [1.11.4] - 2026-09-18
 
 ### Fixed